+++ index.d.ts
@@ -8,7 +8,7 @@
 export function parse(source: string, opts?: ParserOptions): DocumentFragment;
 export function bind(target: DocumentFragment | HTMLElement | ShadowRoot, opts?: ParserOptions): (source: string) => Generator<HTMLElement, undefined, undefined>;
 export function figure(source: DocumentFragment | HTMLElement | ShadowRoot): void;
-export function footnote(source: DocumentFragment | HTMLElement | ShadowRoot, targets: { annotation: HTMLOListElement; authority: HTMLOListElement; }): void;
+export function footnote(source: DocumentFragment | HTMLElement | ShadowRoot, targets: { annotation: HTMLOListElement; reference: HTMLOListElement; }): void;
 export function render(target: HTMLElement, opts?: RenderingOptions): void;
 export function toc(source: DocumentFragment | HTMLElement | ShadowRoot): HTMLUListElement;
 export function info(source: DocumentFragment | HTMLElement | ShadowRoot): Info;
@@ -17,7 +17,7 @@ export interface ParserOptions {
   figure?: boolean;
   footnote?: {
     annotation: HTMLOListElement;
-    authority: HTMLOListElement;
+    reference: HTMLOListElement;
   };
 }