+++ src/parser/inline/reference.test.ts
@@ -1,10 +1,10 @@
-import { authority } from './authority';
+import { reference } from './reference';
 import { some } from '../../combinator';
 import { inspect } from '../../debug.test';
 
-describe('Unit: parser/inline/authority', () => {
-  describe('authority', () => {
-    const parser = some(authority);
+describe('Unit: parser/inline/reference', () => {
+  describe('reference', () => {
+    const parser = some(reference);
 
     it('invalid', () => {
       assert.deepStrictEqual(inspect(parser('')), undefined);
@@ -27,18 +27,18 @@ describe('Unit: parser/inline/authority', () => {
     });
 
     it('basic', () => {
-      assert.deepStrictEqual(inspect(parser('[[a]]')), [['<sup class="authority">a</sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[a ]]')), [['<sup class="authority">a</sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[ab]]')), [['<sup class="authority">ab</sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[a b]]')), [['<sup class="authority">a b</sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[a]]')), [['<sup class="reference">a</sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[a ]]')), [['<sup class="reference">a</sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[ab]]')), [['<sup class="reference">ab</sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[a b]]')), [['<sup class="reference">a b</sup>'], '']);
     });
 
     it('nest', () => {
-      assert.deepStrictEqual(inspect(parser('[["]]')), [['<sup class="authority">"</sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[<a>]]')), [['<sup class="authority"><span class="invalid" data-invalid-syntax="html" data-invalid-type="syntax">&lt;a&gt;</span></sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[`a`]]')), [['<sup class="authority"><code data-src="`a`">a</code></sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[[a]]]')), [['<sup class="authority">[a]</sup>'], '']);
-      assert.deepStrictEqual(inspect(parser('[[@a]]')), [['<sup class="authority"><a class="account" rel="noopener">@a</a></sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[["]]')), [['<sup class="reference">"</sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[<a>]]')), [['<sup class="reference"><span class="invalid" data-invalid-syntax="html" data-invalid-type="syntax">&lt;a&gt;</span></sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[`a`]]')), [['<sup class="reference"><code data-src="`a`">a</code></sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[[a]]]')), [['<sup class="reference">[a]</sup>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[@a]]')), [['<sup class="reference"><a class="account" rel="noopener">@a</a></sup>'], '']);
     });
 
   });