+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.129.0
 
+- Change the syntax name from authority to reference.
 - Change numbering of figures not to inherit the label formats specified by themselves.
 - Fix numbering of figures to have consistent increment patterns.